Install rvcheck 0.1.8 for ggtree
This is to avoid the error "Error: object ‘get_aes_var’ is not exported
by 'namespace:rvcheck'" when installing ggtree.

The latest version of rvcheck (0.2.0) doesn't provide the get_aes_var
function, but the ggtree version, for this Bioconductor version,
requires get_aes_var. Installing rvcheck 0.1.8 seemed the easiest solution.